init aave perp

+++ test/pool/Setup.t.sol
@@ -24,7 +24,7 @@ contract TestPool is Test {
         currency0 = new MockERC20("currency0","currency0",18);
         currency1 = new MockERC20("currency1","currency1",18);
 
-        if (uint160(address(currency0)) < uint160(address(currency1))) {
+        if (address(currency0) < address(currency1)) {
             (currency0, currency1) = (currency1, currency0);
         }
